Set nextJsUrl to be string
It is annoying for the developer experience to force the string to start with `http://` or `https://` as the dev may already have an env variable with the base Url.

I find it pretty useful to use an environment variable so I can change easily between `http://localhost:3000` and the actual value in test and prod env.
```tsx
   <SolitoImageProvider
      nextJsURL={process.env.NEXT_PUBLIC_ROOT_URL}
      ...
    />
```
Forcing the string to start with either https or http means I would need to check for environment in the code of the provider. I would prefer not.